+++ static/app/views/insights/pages/frontend/frontendOverviewTable.tsx
@@ -21,8 +21,8 @@ import {StarredSegmentCell} from 'sentry/views/insights/common/components/tableC
 import {QueryParameterNames} from 'sentry/views/insights/common/views/queryParameters';
 import {DataTitles} from 'sentry/views/insights/common/views/spans/types';
 import {StyledIconStar} from 'sentry/views/insights/pages/backend/backendTable';
-import {getSpanOpFromQuery} from 'sentry/views/insights/pages/frontend/newFrontendOverviewPage';
 import {SPAN_OP_QUERY_PARAM} from 'sentry/views/insights/pages/frontend/settings';
+import {getSpanOpFromQuery} from 'sentry/views/insights/pages/frontend/utils/pageSpanOp';
 import {TransactionCell} from 'sentry/views/insights/pages/transactionCell';
 import type {SpanResponse} from 'sentry/views/insights/types';
